% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simpleRslopes.R
\name{plot.simpleRslopes}
\alias{plot.simpleRslopes}
\title{S3 Plot Method for \code{\link{simpleRslopes}} Objects}
\usage{
\method{plot}{simpleRslopes}(x, one_plot = TRUE, ...)
}
\arguments{
\item{x}{\code{\link{simpleRslopes}} object}

\item{one_plot}{Boolean.
Should one plot be printed?
Defaults to \code{TRUE}}

\item{...}{Additional arguments.
No additional arguments available at this time}
}
\value{
Returns a list containing each plot for the simple slopes
}
\description{
Plot method for \code{\link[simpleRslopes]{simple_slopes}}
}
\examples{
# Generate data
df <- data.frame(
  y = rnorm(100),
  x1 = rnorm(100),
  x2 = rnorm(100)
)

# Estimate linear model
fit <- lm(y ~ x1 * x2, data = df)

# Estimate simple slopes
result <- simple_slopes(
  fit = fit,
  variable1name = "x1", # must be name as it appears in 'fit' object
  variable2name = "x2", # must be name as it appears in 'fit' object
  plot = FALSE # only FALSE for CRAN checks
)

# Plot simple slopes
plot(result)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
