% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.simpleRslopes.R
\name{print.simpleRslopes}
\alias{print.simpleRslopes}
\title{S3 Print Method for \code{\link{simpleRslopes}} Objects}
\usage{
\method{print}{simpleRslopes}(x, ...)
}
\arguments{
\item{x}{\code{\link{simpleRslopes}} object}

\item{...}{Additional arguments.
No additional arguments available at this time}
}
\value{
Returns a results for the simple slopes
}
\description{
Print method for \code{\link[simpleRslopes]{simple_slopes}}
}
\examples{
# Generate data
df <- data.frame(
  y = rnorm(100),
  x1 = rnorm(100),
  x2 = rnorm(100)
)

# Estimate linear model
fit <- lm(y ~ x1 * x2, data = df)

# Estimate simple slopes
simple_slopes(
  fit = fit,
  variable1name = "x1", # must be name as it appears in 'fit' object
  variable2name = "x2", # must be name as it appears in 'fit' object
  plot = FALSE # only FALSE for CRAN checks
)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
