% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_slopes.R
\name{simple_slopes}
\alias{simple_slopes}
\title{Simple Slopes Analysis}
\usage{
simple_slopes(
  fit,
  variable1name,
  variable1type = c("dichotomous", "continuous"),
  variable1values = NULL,
  variable2name,
  variable2type = c("dichotomous", "continuous"),
  variable2values = NULL
)
}
\arguments{
\item{fit}{\code{"lm"} object.
Object from a model fit using a regression model}

\item{variable1name}{Character.
Name of the first variable in interaction from \code{fit}}

\item{variable1type}{Character.
Category type of the first variable in interaction from \code{fit}}

\item{variable1values}{Numeric.
Values to be used when first variable is the moderator.
If \code{variable1type = "dichotomous"}, then
defaults to values 0 and 1.
If \code{variable1type = "continuous"}, then
defaults to values -1 SD, M, and +1 SD}

\item{variable2name}{Character.
Name of the second variable in interaction from \code{fit}}

\item{variable2type}{Character.
Category type of the second variable in interaction from \code{fit}}

\item{variable2values}{Numeric.
Values to be used when second variable is the moderator.
If \code{variable2type = "dichotomous"}, then
defaults to values 0 and 1.
If \code{variable2type = "continuous"}, then
defaults to values -1 SD, M, and +1 SD}
}
\value{
Returns a list containing:

\item{results}{Data frame with variable, moderator, values of moderator
betas of simple slopes, \emph{t} values, \emph{p}-values, and significance level}

\item{parameters}{A list containing degrees of freedom (\code{df}),
betas (from \code{fit}), interaction name, variance-covariance matrix
of the variables involved, and data from \code{fit}}
}
\description{
Computes simple slopes for objects fit with
the class \code{"lm"}
}
\examples{
# Generate data
df <- data.frame(
  y = rnorm(100),
  x1 = rnorm(100),
  x2 = rnorm(100)
)

# Estimate linear model
fit <- lm(y ~ x1 * x2, data = df)

# Estimate simple slopes
simple_slopes(
  fit = fit,
  variable1name = "x1", # must be name as it appears in 'fit' object
  variable1type = "continuous",
  variable2name = "x2", # must be name as it appears in 'fit' object
  variable2type = "continuous"
)

}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
