% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleRslopes.R
\docType{package}
\name{simpleRslopes-package}
\alias{simpleRslopes}
\alias{simpleRslopes-package}
\title{simpleRslopes--package}
\description{
Computes simple slopes (and intercepts) for most
regression-based models in R. Provides straightforward output,
interpretations, and plots of simple slopes to aid in interpretability.
Follows the computation of simple slopes provided by Preacher,
Curran, and Bauer (2006).
}
\references{
Preacher, K. J., Curran, P. J., & Bauer, D. J. (2006).
Computational tools for probing interactions in multiple linear regression, multilevel modeling, and latent curve analysis.
\emph{Journal of Educational and Behavioral Statistics}, \emph{31}(4), 437-448.
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
