% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_slopes.R
\name{simple_slopes}
\alias{simple_slopes}
\title{Simple Slopes Analysis}
\usage{
simple_slopes(
  fit,
  variable1name,
  variable1values = NULL,
  variable2name,
  variable2values = NULL,
  plot_slopes = TRUE
)
}
\arguments{
\item{fit}{\code{"lm"} object.
Object from a model fit using a regression model}

\item{variable1name}{Character.
Name of the first variable in interaction from \code{fit}}

\item{variable1values}{Numeric.
Values to be used when first variable is the moderator.
If \code{variable1type = "dichotomous"}, then
defaults to values 0 and 1.
If \code{variable1type = "continuous"}, then
defaults to values -1 SD, M, and +1 SD}

\item{variable2name}{Character.
Name of the second variable in interaction from \code{fit}}

\item{variable2values}{Numeric.
Values to be used when second variable is the moderator.
If \code{variable2type = "dichotomous"}, then
defaults to values 0 and 1.
If \code{variable2type = "continuous"}, then
defaults to values -1 SD, M, and +1 SD}

\item{plot_slopes}{Boolean.
Should simple slopes be plotted?
Defaults to \code{TRUE}}
}
\value{
Returns a list containing:

\item{fit}{Returns \code{fit} object back}

\item{results}{Data frame with variable, moderator, values of moderator
betas of simple slopes, \emph{t} values, \emph{p}-values, and significance level}

\item{parameters}{A list containing degrees of freedom (\code{df}),
betas (from \code{fit}), interaction name, variance-covariance matrix
of the variables involved, and data from \code{fit}}

\item{plot_args}{Internal use only. Arguments for plots}
}
\description{
Computes simple slopes for objects fit with
the class \code{"lm"}
}
\examples{
# Generate data
df <- data.frame(
  y = rnorm(100),
  x1 = rnorm(100),
  x2 = rnorm(100)
)

# Estimate linear model
fit <- lm(y ~ x1 * x2, data = df)

# Estimate simple slopes
simple_slopes(
  fit = fit,
  variable1name = "x1", # must be name as it appears in 'fit' object
  variable2name = "x2", # must be name as it appears in 'fit' object
  plot = TRUE
)

}
\references{
Preacher, K. J., Curran, P. J., & Bauer, D. J. (2006).
Computational tools for probing interactions in multiple linear regression, multilevel modeling, and latent curve analysis.
\emph{Journal of Educational and Behavioral Statistics}, \emph{31}(4), 437-448.
}
\author{
Alexander Christensen <alexpaulchristensen@gmail.com>
}
